refactor server routes and fix navigate bug
